Add sidebars
+ improve UI
+ refactor renderer.js -> wordToHtml.js
