+++ main.js
@@ -1,4 +1,5 @@
 const electron = require('electron')
+const {Menu} = require('electron')
 // Module to control application life.
 const app = electron.app
 // Module to create native browser window.
@@ -14,7 +15,12 @@ let mainWindow
 
 function createWindow () {
   // Create the browser window.
-  mainWindow = new BrowserWindow({width: 800, height: 600})
+  mainWindow = new BrowserWindow({
+      width: 800,
+      height: 600,
+      minWidth: 600,
+      minHeight: 400
+      })
 
   // and load the index.html of the app.
   mainWindow.loadURL(url.format({
